1. Couple of ToDos 2. Agent Monitor
