+++ src/Microsoft.VisualStudio.Services.Agent/ProcessInvoker.cs
@@ -2,10 +2,8 @@ using Microsoft.VisualStudio.Services.Agent.Util;
 using System;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
-using System.ComponentModel;
 using System.Diagnostics;
 using System.IO;
-using System.Linq;
 using System.Runtime.InteropServices;
 using System.Text;
 using System.Threading;
@@ -24,7 +22,8 @@ namespace Microsoft.VisualStudio.Services.Agent
             string fileName,
             string arguments,
             IDictionary<string, string> environment,
-            CancellationToken cancellationToken);
+            CancellationToken cancellationToken,
+            bool waitForExit = true);
 
         Task<int> ExecuteAsync(
             string workingDirectory,
@@ -32,7 +31,8 @@ namespace Microsoft.VisualStudio.Services.Agent
             string arguments,
             IDictionary<string, string> environment,
             bool requireExitCodeZero,
-            CancellationToken cancellationToken);
+            CancellationToken cancellationToken,
+            bool waitForExit = true);
 
         Task<int> ExecuteAsync(
             string workingDirectory,
@@ -41,7 +41,8 @@ namespace Microsoft.VisualStudio.Services.Agent
             IDictionary<string, string> environment,
             bool requireExitCodeZero,
             Encoding outputEncoding,
-            CancellationToken cancellationToken);
+            CancellationToken cancellationToken,
+            bool waitForExit = true);
 
         Task<int> ExecuteAsync(
             string workingDirectory,
@@ -51,7 +52,8 @@ namespace Microsoft.VisualStudio.Services.Agent
             bool requireExitCodeZero,
             Encoding outputEncoding,
             bool killProcessOnCancel,
-            CancellationToken cancellationToken);
+            CancellationToken cancellationToken,
+            bool waitForExit = true);
     }
 
     // The implementation of the process invoker does not hook up DataReceivedEvent and ErrorReceivedEvent of Process,
@@ -82,7 +84,8 @@ namespace Microsoft.VisualStudio.Services.Agent
             string fileName,
             string arguments,
             IDictionary<string, string> environment,
-            CancellationToken cancellationToken)
+            CancellationToken cancellationToken,
+            bool waitForExit=true)
         {
             return ExecuteAsync(
                 workingDirectory: workingDirectory,
@@ -90,7 +93,8 @@ namespace Microsoft.VisualStudio.Services.Agent
                 arguments: arguments,
                 environment: environment,
                 requireExitCodeZero: false,
-                cancellationToken: cancellationToken);
+                cancellationToken: cancellationToken,
+                waitForExit: waitForExit);
         }
 
         public Task<int> ExecuteAsync(
@@ -99,7 +103,8 @@ namespace Microsoft.VisualStudio.Services.Agent
             string arguments,
             IDictionary<string, string> environment,
             bool requireExitCodeZero,
-            CancellationToken cancellationToken)
+            CancellationToken cancellationToken,
+            bool waitForExit=true)
         {
             return ExecuteAsync(
                 workingDirectory: workingDirectory,
@@ -108,7 +113,8 @@ namespace Microsoft.VisualStudio.Services.Agent
                 environment: environment,
                 requireExitCodeZero: requireExitCodeZero,
                 outputEncoding: null,
-                cancellationToken: cancellationToken);
+                cancellationToken: cancellationToken,
+                waitForExit: waitForExit);
         }
 
         public Task<int> ExecuteAsync(
@@ -118,7 +124,8 @@ namespace Microsoft.VisualStudio.Services.Agent
             IDictionary<string, string> environment,
             bool requireExitCodeZero,
             Encoding outputEncoding,
-            CancellationToken cancellationToken)
+            CancellationToken cancellationToken,
+            bool waitForExit=true)
         {
             return ExecuteAsync(
                 workingDirectory: workingDirectory,
@@ -128,7 +135,8 @@ namespace Microsoft.VisualStudio.Services.Agent
                 requireExitCodeZero: requireExitCodeZero,
                 outputEncoding: outputEncoding,
                 killProcessOnCancel: false,
-                cancellationToken: cancellationToken);
+                cancellationToken: cancellationToken,
+                waitForExit: waitForExit);
         }
 
         public async Task<int> ExecuteAsync(
@@ -139,7 +147,8 @@ namespace Microsoft.VisualStudio.Services.Agent
             bool requireExitCodeZero,
             Encoding outputEncoding,
             bool killProcessOnCancel,
-            CancellationToken cancellationToken)
+            CancellationToken cancellationToken,
+            bool waitForExit=true)
         {
             ArgUtil.Null(_proc, nameof(_proc));
             ArgUtil.NotNullOrEmpty(fileName, nameof(fileName));
@@ -213,6 +222,13 @@ namespace Microsoft.VisualStudio.Services.Agent
                 _proc.StandardInput.Dispose();
             }
 
+            if(!waitForExit)
+            {
+                //just give some time for the AgentService.exe to warm up
+                Thread.Sleep(5*1000);
+                return 0;
+            }
+            
             // Start the standard error notifications, if appropriate.
             if (_proc.StartInfo.RedirectStandardError)
             {
